fix(secure,bridge,cluster): close the six defects the 2026-09-06 audit found
Every fix mirrors matter.js HEAD and carries a guard that is red with the
fix reverted (each one observed) and green with it restored:

- AccessControl.ACL: the per-fabric cap counted entries by the FabricIndex
  the client put on the wire, then stamped the writer's fabric on all of
  them — twenty entries tagged with a foreign index passed a limit of four.
  The count is the list as stored (AccessControlServer.ts:186-189 reaches
  the same number after its fabric-scoped write has stamped the fabric).
- Privacy: the mask covered one AES block, the region is up to 20 bytes
  (MessagePrivacy.ts:53-56 runs CTR over all of it); PrivacyKeystream now
  yields whole blocks for the region. A unicast frame with the P bit is
  dropped, as matter.js ExchangeManager.ts:220-224 and chip do.
- Nonce: Header.NonceSecurityFlags returns the received byte, reserved
  bits included (MessageCodec.ts:45); Session.Decrypt builds the nonce
  from the session's peer id, never the header's (NodeSession.ts:187).
- PASE MRP parameters: idle/active retransmit timeouts are uint32 on the
  wire (PaseMessages.ts:25-29) and were wrapped into 16 bits.
- sigma.Responder: SetResumptionStore / SetSessionParameters take r.mu.
- subscription.Manager.Start is idempotent, as its doc said.

Consumer note: spake2.MRPParameters.IdleRetransTimeoutMs and
ActiveRetransTimeoutMs are *uint32 now.

+++ bridge/pase_mrp_params_test.go
@@ -101,7 +101,7 @@ func TestPaseAdapter_EmitsResponderMRPParams(t *testing.T) {
 	a := NewPaseAdapter(nil)
 	a.SetPBKDFParams(1000, []byte("SPAKE2P Key Salt"), 1)
 	a.randomSource = func() [spake2.PBKDFRandomSize]byte { return [spake2.PBKDFRandomSize]byte{0x11} }
-	idle, active, thresh := uint16(500), uint16(300), uint16(4000)
+	idle, active, thresh := uint32(500), uint32(300), uint16(4000)
 	a.SetResponderMRPParams(&spake2.MRPParameters{
 		IdleRetransTimeoutMs:   &idle,
 		ActiveRetransTimeoutMs: &active,
